% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaforGUI.R
\name{metaforGUI}
\alias{metaforGUI}
\title{metaforGUI}
\usage{
metaforGUI()
}
\description{
Graphical User Interface (GUI) for the R metafor Package
}
\examples{
# Simply run the main function to bring up the GUI.
metaforGUI()
# Loading data set, selecting variables and running analysis are run from the GUI.
}
\references{
[1] <metafor package>
}

